+++ fftw/api/print-plan.c
@@ -18,11 +18,11 @@
  *
  */
 
-#include "api.h"
+#include "api/api.h"
 
 char *X(sprint_plan)(const X(plan) p)
 {
-     int cnt;
+     size_t cnt;
      char *s;
      plan *pln = p->pln;
